backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "fastapi-mail>=1.5.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.5",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.42",
    "uvicorn>=0.35.0",
    "google-generativeai>=0.7.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.3.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
]
